+++ src/compile.task.ts
@@ -1,11 +1,11 @@
 import {Task} from './task/task.js';
-import {compile} from './compile/compile.js';
+import {compileSourceDirectory} from './compile/compileSourceDirectory.js';
 import {cleanBuild} from './project/clean.js';
 
 export const task: Task = {
 	description: 'compiles all files to the build directory',
 	run: async ({log}) => {
 		await cleanBuild(log);
-		await compile(log);
+		await compileSourceDirectory(log);
 	},
 };