+++ src/compile/tsHelpers.ts
@@ -1,5 +1,5 @@
 import ts from 'typescript';
-import {dirname} from 'path';
+import {join, dirname, resolve} from 'path';
 
 import {black, bgRed} from '../colors/terminal.js';
 import {Logger} from '../utils/log.js';
@@ -18,11 +18,26 @@ export interface TsConfig {
 	compileOnSave?: boolean;
 }
 
+const tsconfigCache: Map<string, TsConfig> = new Map();
+
+// TODO This is pretty slow.
+// (10ms last I measured, might seem small but you can do a LOT of work in 10ms and it's *blocking*)
+// Caching helps but maybe we should just import the JSON, at least when only using swc?
+// Also we don't currently watch for changes, but could eventually,
+// way down the line when that's the biggest issue to address!
 export const loadTsconfig = (
 	log: Logger,
 	tsconfigPath?: string,
 	basePath = tsconfigPath ? dirname(tsconfigPath) : process.cwd(),
+	forceReload = false,
 ): TsConfig => {
+	// create a canonical cache key that can accept multiple variations
+	const cacheKey = join(resolve(basePath), tsconfigPath || 'tsconfig.json');
+	if (!forceReload) {
+		const cachedTsconfig = tsconfigCache.get(cacheKey);
+		if (cachedTsconfig) return cachedTsconfig;
+	}
+
 	if (!tsconfigPath) {
 		const searchPath = tsconfigPath || basePath;
 		tsconfigPath = ts.findConfigFile(searchPath, ts.sys.fileExists);
@@ -34,18 +49,20 @@ export const loadTsconfig = (
 	const readResult = ts.readConfigFile(tsconfigPath, ts.sys.readFile);
 	if (readResult.error) logTsDiagnostics([readResult.error], log);
 
-	const config: TsConfig = readResult.config;
-	if (!config) throw Error(`Unable to read tsconfig from ${tsconfigPath}`);
+	const tsconfig: TsConfig = readResult.config;
+	if (!tsconfig) throw Error(`Unable to read tsconfig from ${tsconfigPath}`);
 
-	const rawCompilerOptions = config.compilerOptions;
+	const rawCompilerOptions = tsconfig.compilerOptions;
 	const convertResult = ts.convertCompilerOptionsFromJson(rawCompilerOptions, basePath);
 	if (convertResult.errors) logTsDiagnostics(convertResult.errors, log);
 
-	return {
-		...config,
-		compilerOptions: convertResult.options,
-		rawCompilerOptions,
-	};
+	// the TypeScript API generally uses the converted options,
+	// but sometimes it's better to have the plain JSON versions so we store both
+	tsconfig.compilerOptions = convertResult.options;
+	tsconfig.rawCompilerOptions = rawCompilerOptions;
+	tsconfigCache.set(cacheKey, tsconfig);
+
+	return tsconfig;
 };
 
 export const logTsDiagnostics = (diagnostics: ReadonlyArray<ts.Diagnostic>, log: Logger): void => {