+++ src/project/rollup-plugin-diagnostics.ts
@@ -5,7 +5,7 @@ import {SystemLogger} from '../utils/log.js';
 import {printKeyValue, printMs, printPath} from '../utils/print.js';
 import {createStopwatch} from '../utils/time.js';
 
-const name = 'diagnostics';
+export const name = 'diagnostics';
 
 const tag = (s: string) => s; // maybe color this